+++ config/checkstyle-resources-suppressions.xml
@@ -79,10 +79,10 @@
   <!-- intentional problem for testing -->
   <suppress checks="PackageDeclarationCheck"
     files="uncommentedmain[\\/]InputUncommentedMainBeginTreePackage2\.java"/>
-  
+
   <!-- intentional problem for testing -->
   <suppress checks="RegexpSingleline"
-    files="javadoctype[\\/]InputJavadocTypeWithBlockComment\.java"/>          
+    files="javadoctype[\\/]InputJavadocTypeWithBlockComment\.java"/>
 
   <!-- intentional problem for testing -->
   <suppress checks="OuterTypeFilename"
@@ -113,17 +113,17 @@
             files="[\\/]javadocparagraph[\\/]InputJavadocParagraphCheck1\.java"/>
   <suppress id="noTrailingWhitespace"
             files="[\\/]javadocparagraph[\\/]InputJavadocParagraphIncorrect4\.java"/>
-  
+
   <!-- Input file contains trailing whitespace to kill pitest mutation -->
-  <suppress id="noTrailingWhitespace" 
+  <suppress id="noTrailingWhitespace"
     files="separatorwrap[\\/]InputSeparatorWrapForTestTrailingWhitespace\.java"/>
-  <suppress id="noTrailingWhitespace" 
+  <suppress id="noTrailingWhitespace"
     files="nonemptyatclausedescription[\\/]InputNonEmptyAtclauseDescriptionOne\.java"/>
-  <suppress id="noTrailingWhitespace" 
+  <suppress id="noTrailingWhitespace"
     files="requireemptylinebeforeblocktaggroup[\\/]InputRequireEmptyLineBeforeBlockTagGroupIncorrect\.java"/>
-  <suppress id="noTrailingWhitespace" 
+  <suppress id="noTrailingWhitespace"
     files="requireemptylinebeforeblocktaggroup[\\/]InputRequireEmptyLineBeforeBlockTagGroupCorrect\.java"/>
-  <suppress id="noTrailingWhitespace" 
+  <suppress id="noTrailingWhitespace"
     files="javadoctagcontinuationindentation[\\/]InputJavadocTagContinuationIndentationDescription\.java"/>
 
   <!-- intentional problem for testing -->
@@ -578,8 +578,6 @@
     files="[\\/]test[\\/]resources[\\/]com[\\/]puppycrawl[\\/]tools[\\/]checkstyle[\\/]filters[\\/]suppressioncommentfilter[\\/]InputSuppressionCommentFilter5\.java"/>
   <suppress checks="FileLength"
     files="[\\/]test[\\/]resources[\\/]com[\\/]puppycrawl[\\/]tools[\\/]checkstyle[\\/]main[\\/]InputMainComplexityOverflow\.java"/>
-  <suppress checks="FileLength"
-    files="[\\/]test[\\/]resources-noncompilable[\\/]com[\\/]puppycrawl[\\/]tools[\\/]checkstyle[\\/]checks[\\/]blocks[\\/]needbraces[\\/]InputNeedBracesTestSwitchExpression\.java"/>
   <suppress checks="FileLength"
     files="[\\/]test[\\/]resources[\\/]com[\\/]puppycrawl[\\/]tools[\\/]checkstyle[\\/]checks[\\/]metrics[\\/]javancss[\\/]InputJavaNCSSRecordsAndCompactCtors\.java"/>
   <suppress checks="FileLength"
